requirements.txt
Django==4.0.4
dlib==19.24.0
face_recognition==1.3.0
imutils==0.5.4
matplotlib==3.4.3
numpy==1.21.3
opencv_python==4.5.5.62
pandas==1.3.4
Pillow==9.1.1
